requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.3
tzdata>=2024.1
anthropic>=0.40.0
pydantic>=2.0.0
